requirements.txt
altair==5.5.0
anthropic==0.46.0
bokeh==3.6.3
matplotlib==3.10.0
numpy==2.2.3
openai==1.63.2
pandas==2.2.3
plotly==6.0.0
seaborn==0.13.2
streamlit==1.42.2
pillow==11.1.0
